Fix config formatting
